refactor(timestampms): move to seismic crate
Another small PR part of refactoring to minimize our forks and maintain as much code as we can in seismic related crates (which we plan to eventually bring into our monorepo).
