+++ crates/bytecode/src/opcode.rs
@@ -466,9 +466,6 @@ opcodes! {
     0x49 => BLOBHASH     => stack_io(1, 1);
     0x4A => BLOBBASEFEE  => stack_io(0, 1);
 
-    // Modification: Added TIMESTAMPMS opcode
-    0x4B => TIMESTAMPMS => stack_io(0, 1);
-
     // 0x4B
     // 0x4C
     // 0x4D
@@ -723,7 +720,7 @@ mod tests {
         for _ in OPCODE_INFO.into_iter().flatten() {
             opcode_num += 1;
         }
-        assert_eq!(opcode_num, 151);
+        assert_eq!(opcode_num, 150);
     }
 
     #[test]
