+++ crates/seismic/src/instructions.rs
@@ -1,3 +1,4 @@
+pub mod block_info;
 pub mod confidential_storage;
 pub mod instruction_provider;
 pub mod seismic_host;
